Completed adding clickHandler on Todo.js.
