+++ src/App.js
@@ -1,24 +1,48 @@
 import React from 'react';
 import ReactDom from 'react-dom';
-
-import Todo from './components/TodoComponents/Todo.js'
+import './components/TodoComponents/Todo.css'
 import TodoList from './components/TodoComponents/TodoList.js'
+
+const TodoData = [
+  {
+    task: 'Organize Garage',
+    id: 1528817077286,
+    completed: false
+  },
+  {
+    task: 'Bake Cookies',
+    id: 1528817084358,
+    completed: false
+  }
+];
 
 class App extends React.Component {
   constructor() {
     super();
     this.state = {
-      toDo: Todo
+      todo: TodoData
     }
   }
-  // you will need a place to store your state in this component.
-  // design `App` to be the parent component of your application.
-  // this component is going to take care of state, and any change handlers you need to work with your state
+  
+  clickHandler = id => {
+    this.setState({
+      todo: this.state.todo.map(item => {
+        if (item.id === id) {
+          return {
+            ...item,
+            completed: !item.completed
+          };
+        } else {
+          return item;
+        }
+      })
+    })
+  }
   render() {
     return (
       <div>
         <h2>Welcome to your Todo App!</h2>
-        <TodoList {...this.state} />
+        <TodoList {...this.state} clickHandler={this.clickHandler}/>
       </div>
     );
   }
