+++ src/components/TodoComponents/Todo.css
@@ -1,0 +1,3 @@
+.item.completed {
+    text-decoration: line-through;
+  }
