+++ src/components/TodoComponents/Todo.js
@@ -1,16 +1,14 @@
 import React from 'react';
 
-const TodoData = [
-  {
-    task: 'Organize Garage',
-    id: 1528817077286,
-    completed: false
-  },
-  {
-    task: 'Bake Cookies',
-    id: 1528817084358,
-    completed: false
-  }
-];
 
-export default TodoData;
+const Todo = props => {
+  return(
+    <div 
+      className={`item${props.item.completed ? " completed" : ""}`}
+      onClick={() => props.clickHandler(props.item.id)}>
+      {props.item.task}
+    </div>
+  )
+} 
+
+export default Todo;
